Optimize indexing by recycling database decouments
Before indexing, gather all symbols that will be removed.  Then, when
adding a new symbol, reuse the documents from that list, or really add
a new document to the db if that list becomes empty.  After indexing,
remove all remaining documents that are in that list from the db.

* bdx/index.py (SymbolIndex.delete_file): Remove.
(SymbolIndex.replace_symbol):
(SymbolIndex.delete_document):
(SymbolIndex.get_docs_for_path): New.
(_index_single_file): Add `outdated_documents' argument that contains
the list of documents to recycle from the db.  Use it to replace the
symbols in the database when possible instead of always adding them.
(_WorkerPool.__init__): Add a `files_to_delete' argument that is used
to gather list of documents to recycle from the database.
(_WorkerPool._worker): Gather list of documents to recycle.
(index_binary_directory): Do not delete any documents from the
database.  Instead, pass the list of files to remove to the worker, so
it can gather documents to recycle and delete them later.