Speed up startup by deferring the opening of all shards for an index
Do not open all shards for a SymbolIndex if we want to just get the
binary dir of it.

* bdx/cli.py (guess_index_path_for_directory): Do not open the
SymbolIndex, instead call the new static `get_binary_dir' function to
try to get the binary directory for the index at given path.
* bdx/index.py (DB_SUBDIR): New helper var.
(SymbolIndex.open): Use it.
(SymbolIndex.open_shard): Use it and remove the `readonly' argument as
shards should only be used for writing.
(SymbolIndex.get_binary_dir): New staticmethod.